+++ mflike/MFLike.yaml
@@ -54,179 +54,13 @@ data:
     - experiments: [LAT_145, LAT_225]
     - experiments: [LAT_225, LAT_225]
 
-# Parameters to build a top-hat band:
-# - nsteps sets the number of frequencies used in the band integration
-# - bandwidth sets the relative width of the band wrt the central frequency
-#   the central frequency of each band is set from the bands stored in the sacc file
-#     - with nstep: 1, bandwidth must be 0
-#     Dirac delta bandpass, no band integration
-#     useful if you don't want to do band integration
-#     when the bandpasses in the sacc file are multifrequency
-#     the freq used is the effective frequency from the bandpass
-#     - if nstep > 1,  bandwidth must be > 1
-#     bandwidth can be a list if you want a different width for each band
-#     e.g. bandwidth: [0.3,0.2,0.3] for 3 bands
-# when top_hat_band is a null dict: no top-hat band is built and
-# bandpasses read from sacc file. Band integration is performed accordingly
-# (if bandpass in sacc is a single freq, no band integration)
-# the default is to read bandpasses from file, to build top-hat uncomment the
-# parameters of the block!
-# Bandpass has to be divided by nu^2 if measured with respect to a RJ source
-# the nu^2 factor is set back again in the _bandpass_construction function
-top_hat_band:
-#  nsteps: 1
-#  bandwidth: 0
-
 # uncomment the block to include a systematic template
 # to be read from external file at "rootname"
 # default is systematic_template to be a null dict
 systematics_template:
 #  rootname: "test_template"
 
-foregrounds:
-  normalisation:
-    nu_0: 150.0
-    ell_0: 3000
-    T_CMB: 2.725
-
-  components:
-    tt:
-      - kSZ
-      - tSZ_and_CIB
-      - cibp
-      - dust
-      - radio
-    te:
-      - radio
-      - dust
-    ee:
-      - radio
-      - dust
-
-params:
-  # Foregrounds
-  a_tSZ:
-    prior:
-      min: 3.0
-      max: 3.6
-    proposal: 0.05
-    latex: a_\mathrm{tSZ}
-  a_kSZ:
-    prior:
-      min: 1.4
-      max: 1.8
-    proposal: 0.1
-    latex: a_\mathrm{kSZ}
-  a_p:
-    prior:
-      min: 6.2
-      max: 7.6
-    proposal: 0.075
-    latex: a_p
-  beta_p:
-    prior:
-      min: 1.8
-      max: 2.2
-    proposal: 0.015
-    latex: \beta_p
-  a_c:
-    prior:
-      min: 4.4
-      max: 5.4
-    proposal: 0.12
-    latex: a_c
-  beta_c:
-    prior:
-      min: 2.0
-      max: 2.4
-    proposal: 0.03
-    latex: \beta_c
-  a_s:
-    prior:
-      min: 2.8
-      max: 3.4
-    proposal: 0.01
-    latex: a_s
-  a_gtt:
-    prior:
-      dist: norm
-      loc: 2.79
-      scale: 0.45
-    proposal: 0.4
-    latex: a_\mathrm{dust}^\mathrm{TT}
-  a_gte:
-    prior:
-      dist: norm
-      loc: 0.36
-      scale: 0.04
-    proposal: 0.04
-    latex: a_\mathrm{dust}^\mathrm{TE}
-  a_gee:
-    prior:
-      dist: norm
-      loc: 0.13
-      scale: 0.03
-    proposal: 0.03
-    latex: a_\mathrm{dust}^\mathrm{EE}
-  a_psee:
-    prior:
-      min: 0
-    proposal: 0.05
-    latex: a_\mathrm{ps}^\mathrm{EE}
-  a_pste:
-    prior:
-      min: -1
-      max: 1
-    proposal: 0.05
-    latex: a_\mathrm{ps}^\mathrm{TE}
-  xi:
-    prior:
-      min: 0
-      max: 0.2
-    proposal: 0.05
-    latex: \xi
-  T_d:
-    prior:
-      min:  8.60
-      max: 10.60
-    proposal: 0.6
-    latex: T_d
-  
-  beta_s:               #beta radio
-    value: -2.5
-    latex: \beta_s    
-  alpha_s:              #alpha radio
-    value: 1.0
-    latex: \alpha_s
-  T_effd:               #effective galactic dust temperature
-    value: 19.6
-    latex: T_{\mathrm{dust},\mathrm{eff}}
-  beta_d:               #beta galactic dust
-    value: 1.5
-    latex: \beta_\mathrm{dust}
-  alpha_dT:             #galactic dust ell slope for T
-    value: -0.6
-    latex: \alpha_{\mathrm{dust},T}
-  alpha_dE:             #galactic dust ell slope for E
-    value: -0.4
-    latex: \alpha_{\mathrm{dust},E}
-  alpha_p:              #CIB poisson ell slope
-    value: 1.0
-    latex: \alpha_p
-  alpha_tSZ:             #tSZ ell slope
-    value: 0.0
-    latex: \alpha_{\rm tSZ}
-
-  # Systematics
-  bandint_shift_LAT_93:
-    value: 0
-    latex: \Delta_{\rm band}^{93}
-  bandint_shift_LAT_145:
-    value: 0
-    latex: \Delta_{\rm band}^{145}
-  bandint_shift_LAT_225:
-    value: 0
-    latex: \Delta_{\rm band}^{225}
+  # Systematics and calibration
   calT_LAT_93:
     value: 1
     latex: \mathrm{Cal}_{\rm T}^{93}
