+++ mflike/__init__.py
@@ -1,5 +1,4 @@
 from .mflike import MFLike
-from .theoryforge import TheoryForge
 from importlib.metadata import version, PackageNotFoundError
 
 try:
